cloudrun/requirements.txt
Flask==3.0.3
flask-cors==3.0.10
gunicorn==23.0.0
Werkzeug==3.0.3
langchain>=0.1.0
langchain-community>=0.0.10
torch>=2.0.0
transformers>=4.30.0
ctransformers>=0.2.24  # For local inference
huggingface-hub>=0.16.4
requests==2.31.0
python-dotenv==1.0.0